+++ mayan/apps/dynamic_search/tests/test_search_api.py
@@ -8,8 +8,9 @@ from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
 from ..search_models import SearchModel
 
-from .mixins.search_api_mixins import SearchAPIViewTestMixin
 from .mixins.base import TestSearchObjectSimpleTestMixin
+from .mixins.search_api_mixins import SearchAPIViewTestMixin
+from .mixins.search_model_api_mixins import SearchModelAPIViewTestMixin
 
 
 class SearchAPIViewBackwardCompatilityTestCase(
@@ -192,47 +193,6 @@ class SearchFilterCombinatiomAPITestCase(
         self.assertEqual(
             response.data['count'], 1
         )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-
-class SearchModelAPIViewTestCase(
-    TestSearchObjectSimpleTestMixin, BaseAPITestCase
-):
-    def test_search_models_api_view(self):
-        self._clear_events()
-
-        response = self.get(
-            viewname='rest_api:searchmodel-list', query={'page_size': 50}
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        expected_value = []
-        for search_model in SearchModel.all():
-            search_model_expected_value = {
-                'app_label': search_model.app_label,
-                'model_name': search_model.model_name,
-                'pk': search_model.pk,
-                'search_fields': []
-            }
-
-            for search_field in search_model.search_fields:
-                search_model_expected_value['search_fields'].append(
-                    {
-                        'field_name': search_field.field_name,
-                        'label': search_field.label
-                    }
-                )
-
-            expected_value.append(search_model_expected_value)
-
-        self.assertEqual(
-            response.data['results'], expected_value
-        )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)